+++ lib/index.ts
@@ -8,4 +8,5 @@ export { IntentRequestBuilder } from './factory/IntentRequestBuilder';
 export { LaunchRequestBuilder } from './factory/LaunchRequestBuilder';
 export { SessionEndedRequestBuilder } from './factory/SessionEndedRequestBuilder';
 export { AudioPlayerPauseIntentRequestBuilder, AudioPlayerResumeIntentRequestBuilder } from './factory/AudioIntentRequestBuilder';
+export { AplUserEventRequestBuilder } from './factory/AplUserEventRequestBuilder';
 export { SkillSettings, SequenceItem, ResponseValidator } from './types';