+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 
 ## current master
+* feat: add `renderDocument` parameter used to assert templates in `AplValidator` class.
+* feat: add `AplUserEventRequestBuilder` to provide support for touch events.
 
 ## 2.4.0
 * update libraries